chore: bump @bsull/augurs to 0.6.0
This release includes separate entry points for the various
parts of the library, which should reduce the amount of
WebAssembly that needs to be loaded by the user.

It also fixes build warnings related to circular dependencies.

This commit uses a prerelease version but I'll bump that
once I've tested everything's working.